Number 2/app/requirements.txt
fastapi
uvicorn
opencv-python
numpy
requests
openai
pillow
google-generativeai
python-multipart